This reverts commit 09467fd97563309c43a0419c47f65c57806bdea1.

+++ src/maxdiffusion/models/train.py
@@ -387,8 +387,8 @@ def train(config):
                 noisy_latents = noise_scheduler.add_noise(noise_scheduler_state, latents, new_noise, timesteps)
             else:
                noisy_latents = noise_scheduler.add_noise(noise_scheduler_state, latents, noise, timesteps)
-            noisy_latents = noisy_latents.astype(jnp.bfloat16)
-            encoder_hidden_states = encoder_hidden_states.astype(jnp.bfloat16)
+            #noisy_latents = noisy_latents.astype(jnp.bfloat16)
+            #encoder_hidden_states = encoder_hidden_states.astype(jnp.bfloat16)
             # Predict the noise residual and compute loss
             model_pred = pipeline.unet.apply(
                 {"params": unet_params}, noisy_latents, timesteps, encoder_hidden_states, train=True
